refactor: util 디렉토리 메인 index.ts 생성 및 컴포넌트에서 import 경로 수정
