+++ front/src/pages/balance/Withdrawal.tsx
@@ -1,6 +1,6 @@
 import React, { ChangeEvent, useState } from 'react';
 import TOAST from '@lib/toastify';
-import formatNumber from '@src/common/utils/formatNumber';
+import { formatNumber } from '@common/utils';
 import { withdraw } from '@lib/api';
 
 interface WithdrawalProps {
