+++ front/src/pages/trade/chart/candle/CandleLegend.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
 import { IChartItem, Theme, IUser } from '@src/types';
 import { userAtom } from '@recoil';
-import formatNumber from '@common/utils/formatNumber';
+import { formatNumber } from '@common/utils';
 import {
 	IGraphComponentProps,
 	ICrossLine,
