+++ front/src/pages/trade/order/AskOrderItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { IAskOrderItem } from '@src/types';
-import formatNumber from '@common/utils/formatNumber';
+import { formatNumber } from '@common/utils';
 import ITotalAndMaxAmount from './ITotalAndMaxAmount';
 
 interface IProps {
